i2c-ocores: fu540-c000: support ocore driver for polling mode i2c transfer
Add support to switch i2c-ocore driver to polling mode even if the IRQ
is defined in Device Tree for FU540-c000 chipset.

FU540 chipset has an erratum for I2C, the interrupt-driven i2c transfers
is not working. Even thought the driver requests and assigns an interrupt
from DT, depending on the controller type it sets the I2C flag to
SIFIVE_IRQ_POLL and operates in polling mode for i2C transfers.

Signed-off-by: Sagar Shrikant Kadam <sagar.kadam@sifive.com>