net: phy: fix for ethernet phy VSC8541-01 probe failure
HiFive unleashed A00 board has VSC8541-01 ethernet phy, this phy device
is identified as the Rev B variant in the device identification register.
In order to use this phy in the unmanaged mode, it requires a specific
reset sequence of the logical 0-1-0-1 transition on  NRESET pin [1].

Generally, the bootloader takes care of the phy reset but if it is not
handled in the bootloader the phy device is not recognised in linux.
Linux phylib gives a reset to the phy as per commit bafbdd527d56 ("phylib:
 Add device reset GPIO support") but since the phy VSC8541-01 requires one
more reset the network setup fails.

An additional reset if the phy-id for this phy device is matched fixes this
issue.

Datasheet:
[1] https://www.mouser.com/ds/2/523/Microsemi_VSC8541-01_Datasheet_10496_V40-1148034.pdf

Signed-off-by: Sagar Shrikant Kadam <sagar.kadam@sifive.com>