dts: phy: fix for missing mdio device error for the vsc8541-01 device
HiFive unleashed A00 board has VSC8541-01 ethernet phy, this device is
identified as Revision B device as described in device identification
registers. In order to use this phy in the unmanaged mode, it requires
a specific reset sequence of logical 0-1-0-1 transition on  NRESET pin [1].

The bootloader (fsbl) generates this reset sequence. If the bootloader
does not follow this reset sequence, the linux cadence MACB driver probe
succeeds but the mdio scan fails and is not even able to read the phy
registers. It gives an error message:

[0.733312] libphy: MACB_mii_bus: probed
[0.741730] mdio_bus 10090000.ethernet-ffffffff: MDIO device at address 0 is missing.

Adding the device OUI (Organizationally Unique Identifier) to the phy
device node fixes this problem.

[1]: VSC8541-01 datasheet:
https://www.mouser.com/ds/2/523/Microsemi_VSC8541-01_Datasheet_10496_V40-1148034.pdf

Signed-off-by: Sagar Shrikant Kadam <sagar.kadam@sifive.com>